web/requirements.txt
aiofiles==0.4.0
asyncio-nats-client==0.10.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
envparse==0.2.0
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.4
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
sanic-wtf==0.5.0
sanic_session==0.6.0
ujson==1.35
urllib3==1.25.7
websockets==8.1
